Refactored music streaming code
